Created service to select predefined account
